In-Progress Consensus Creation
I just made this commit so I could add a note on the last one: the create_three_turker_consensus method doesn't currently work. It frequently returns "Couldn't agree!", which means that each of the 3 Turkers said something different and isn't necessarily a problem, but it also marks almost everything as disqualified, even in cases where I think it probably shouldn't (for example, "raceWhite, genderF, ageMiddleAged, voice, Disqualified" shouldn't be a line in this dataframe). I don't have time to finish it right now, but I wanted it known that it doesn't work yet.